Increases mixing of games
Adds to the weightings something to discourage lots of back to back games or breaks. This is because back to back games can leave people tired for the later games and back to back breaks are more boring than intermittent breaks.
Scores to reduce long streaks of playing games, refereeing or free time. Score climbs faster the longer the streak. So a 1 game streak adds 0 score, 2 games = 1 unit, 3=3, 4=6, 5=10...